% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_fonts}
\alias{register_fonts}
\title{Function to register fonts of the TFFF}
\usage{
register_fonts(verbose = FALSE)
}
\arguments{
\item{verbose}{whether or not to display families and their path in the console (FALSE by default)}
}
\description{
Function to register fonts of the TFFF
}
