% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{t_col}
\alias{t_col}
\title{Transparency color - from https://www.dataanalytics.org.uk/make-transparent-colors-in-r/}
\usage{
t_col(color, percent = 50)
}
\arguments{
\item{color}{A color}

\item{percent}{Transparency shade}
}
\value{
A color hex
}
\description{
Transparency color - from https://www.dataanalytics.org.uk/make-transparent-colors-in-r/
}
\examples{
t_col(tfff_colors("Dark Green"),50)
}
