% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{tfff-comma}
\alias{tfff-comma}
\alias{tfff_label_comma}
\alias{tfff_comma}
\alias{tfff_scale_x_comma}
\alias{tfff_scale_y_comma}
\title{Comma formatting}
\usage{
tfff_label_comma(accuracy = 0.1, ...)

tfff_comma(values, accuracy = 0.1, ...)

tfff_scale_x_comma(...)

tfff_scale_y_comma(...)
}
\arguments{
\item{accuracy}{Accuracy of labels (e.g. 0.1 or 0.01)}

\item{...}{Other arguments that are passed to `scales::comma()`, `scales::label_comma()` or `ggplot2::scale_(x|y)_continuous`}

\item{values}{Values to be formatted}
}
\value{
A character vector of comma labels or a ggplot2 scale.
}
\description{
These functions create comma labels.
They are mostly just wrappers around scales::comma(), scales::label_comma() and ggplot2::scale_(x|y)_continuous().
But they do have the extra convenience built in that the scale functions already use comma formatting.
}
