% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar.R
\name{tfff-dollar}
\alias{tfff-dollar}
\alias{tfff_label_dollar}
\alias{tfff_dollar}
\alias{tfff_scale_x_dollar}
\alias{tfff_scale_y_dollar}
\title{Dollar formatting}
\usage{
tfff_label_dollar(values, reference_values = values, uses_suffix = TRUE)

tfff_dollar(values, reference_values = values, uses_suffix = TRUE)

tfff_scale_x_dollar(uses_suffix = FALSE, ...)

tfff_scale_y_dollar(uses_suffix = FALSE, ...)
}
\arguments{
\item{values}{Vector of numbers}

\item{reference_values}{Overarching values (e.g. all values of a category and not just of one county.)}

\item{...}{Other arguments that are passed to `ggplot2::scale_(x|y)_continuous`}

\item{use_suffix}{Should dollar amounts larger than 10000 be abbreviated as 10K? Defaults to TRUE for non-scale functions and to FALSE otherwise.}
}
\value{
A character vector of dollar amounts or a ggplot2 scale.
}
\description{
These functions format numbers like dollar amounts.
They are mostly a wrapper around scales::label_dollar, scales::dollar, and ggplot2::scale_*_continuous.
The major diference is that there are two arguments `values` and `reference_values`.
The latter value can be used to make sure that dollar formatting happens based on all values and not just
of values wrt. one county.
}
