% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfff-colors.R
\name{tfff_colors}
\alias{tfff_colors}
\title{Function to extract TFFF colors as hex codes}
\usage{
tfff_colors(...)
}
\arguments{
\item{...}{color or colors you want to retun}
}
\value{
A vector of colors
}
\description{
tfff_colors() enables you to pull colors directly from the tfff palette.
Choose one of the following options: `Dark Green`, `Light Green`, `Orange`,
`Yellow`, `Blue`, `Red`, `Brown`, `Dark Gray`, `Medium Gray`, `Light Gray`.
}
