% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfff_draft.R
\name{tfff_draft}
\alias{tfff_draft}
\title{TFFF paged template}
\usage{
tfff_draft(...)
}
\arguments{
\item{...}{Params to pagedown::html_paged}
}
\value{
An Rmd paged them
}
\description{
TFFF paged template
}
