% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{tfff_get_basemap}
\alias{tfff_get_basemap}
\title{Function to get an sf object with the TFFF service area (Oregon and Siskiyou
County, California)}
\usage{
tfff_get_basemap()
}
\value{
sf object with Oregon counties and Siskiyou County, California
}
\description{
Function to get an sf object with the TFFF service area (Oregon and Siskiyou
County, California)
}
