% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfff_offices}
\alias{tfff_offices}
\title{Locations of TFFF offices}
\format{
A data frame with 2 rows and 4 variables
\describe{
  \item{city}{city of office location}
  \item{state}{state of office location}
  \item{lon}{longitude}
  \item{lat}{latitude}
  ...
}
}
\source{
\url{http://tfff.org}
}
\usage{
tfff_offices
}
\description{
A dataset containing locations of TFFF offices.
}
\keyword{datasets}
